[RENAME] user 도메인 디렉토리 생성 후 클래스 이동
