+++ team/src/main/java/sansam/team/user/command/dto/JwtToken.java
@@ -1,4 +1,4 @@
-package sansam.team.test.command.dto;
+package sansam.team.user.command.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
