+++ team/src/main/java/sansam/team/user/command/repository/UserRepository.java
@@ -1,11 +1,12 @@
-package sansam.team.test.command.repository;
+package sansam.team.user.command.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sansam.team.test.command.entity.User;
+import sansam.team.user.command.entity.User;
 
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByIdAndPassword(String id, String password);
+    Optional<User> findById(String id);
 
 }
